+++ example/src/components/SliderEntry.js
@@ -1,39 +1,68 @@
 import React, { Component, PropTypes } from 'react';
 import { View, Text, Image, TouchableOpacity } from 'react-native';
+import { ParallaxImage } from 'react-native-snap-carousel';
 import styles from 'example/src/styles/SliderEntry.style';
 
 export default class SliderEntry extends Component {
 
     static propTypes = {
-        title: PropTypes.string.isRequired,
-        subtitle: PropTypes.string,
-        illustration: PropTypes.string,
-        even: PropTypes.bool
+        data: PropTypes.object.isRequired,
+        even: PropTypes.bool,
+        parallax: PropTypes.bool,
+        parallaxProps: PropTypes.object
     };
 
+    get image () {
+        const { data: { illustration }, parallax, parallaxProps, even } = this.props;
+
+        return parallax ? (
+            <ParallaxImage
+              source={{ uri: illustration }}
+              containerStyle={[styles.imageContainer, even ? styles.imageContainerEven : {}]}
+              style={[styles.image, { position: 'relative' }]}
+              parallaxFactor={0.35}
+              showSpinner={true}
+              spinnerColor={even ? 'rgba(255, 255, 255, 0.4)' : 'rgba(0, 0, 0, 0.25)'}
+              {...parallaxProps}
+            />
+        ) : (
+            <Image
+              source={{ uri: illustration }}
+              style={styles.image}
+            />
+        );
+    }
+
     render () {
-        const { title, subtitle, illustration, even } = this.props;
+        const { data: { title, subtitle }, even } = this.props;
 
         const uppercaseTitle = title ? (
-            <Text style={[styles.title, even ? styles.titleEven : {}]} numberOfLines={2}>{ title.toUpperCase() }</Text>
+            <Text
+              style={[styles.title, even ? styles.titleEven : {}]}
+              numberOfLines={2}
+            >
+                { title.toUpperCase() }
+            </Text>
         ) : false;
 
         return (
             <TouchableOpacity
-              activeOpacity={0.7}
+              activeOpacity={1}
               style={styles.slideInnerContainer}
               onPress={() => { alert(`You've clicked '${title}'`); }}
               >
                 <View style={[styles.imageContainer, even ? styles.imageContainerEven : {}]}>
-                    <Image
-                      source={{ uri: illustration }}
-                      style={styles.image}
-                    />
+                    { this.image }
                     <View style={[styles.radiusMask, even ? styles.radiusMaskEven : {}]} />
                 </View>
                 <View style={[styles.textContainer, even ? styles.textContainerEven : {}]}>
                     { uppercaseTitle }
-                    <Text style={[styles.subtitle, even ? styles.subtitleEven : {}]} numberOfLines={2}>{ subtitle }</Text>
+                    <Text
+                      style={[styles.subtitle, even ? styles.subtitleEven : {}]}
+                      numberOfLines={2}
+                    >
+                        { subtitle }
+                    </Text>
                 </View>
             </TouchableOpacity>
         );