+++ src/pagination/Pagination.js
@@ -9,9 +9,9 @@ export default class Pagination extends Component {
     static propTypes = {
         dotsLength: PropTypes.number.isRequired,
         activeDotIndex: PropTypes.number.isRequired,
-        containerStyle: ViewPropTypes.style,
-        dotStyle: ViewPropTypes.style,
-        inactiveDotStyle: ViewPropTypes.style,
+        containerStyle: ViewPropTypes ? ViewPropTypes.style : View.propTypes.style,
+        dotStyle: ViewPropTypes ? ViewPropTypes.style : View.propTypes.style,
+        inactiveDotStyle: ViewPropTypes ? ViewPropTypes.style : View.propTypes.style,
         inactiveDotOpacity: PropTypes.number,
         inactiveDotScale: PropTypes.number
     };
@@ -28,7 +28,7 @@ export default class Pagination extends Component {
 
         let dots = [];
 
-        for (var i = 0; i < dotsLength; i++) {
+        for (let i = 0; i < dotsLength; i++) {
             dots.push(
                 <PaginationDot
                   key={`pagination-dot-${i}`}